+++ src/main/java/com/launchacademy/giantleap/models/FashionItemNotFoundException.java
@@ -1,0 +1,7 @@
+package com.launchacademy.giantleap.models;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class FashionItemNotFoundException extends RuntimeException {};
+
